pyproject.toml
[project]
name = "devopsagent-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.129.0",
    "lancedb>=0.29.2",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langfuse[langchain]>=3.14.4",
    "langgraph>=1.0.8",
    "loguru>=0.7.3",
    "mcp>=1.26.0",
    "ollama>=0.6.1",
    "pandas>=2.2.1,<3.0.0",
    "prompt-toolkit>=3.0.52",
    "pyarrow>=15.0.0",
    "pydantic-settings>=2.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pyyaml>=6.0.1",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "uvicorn[standard]>=0.29.0,<0.35.0",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
devops-agent = "src.cli.main:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
